# api/openapi.yaml
openapi: 3.0.3
info:
  title: TalkFlow API
  description: API for TalkFlow, a smart video/audio call application with AI-powered notes.
  version: 1.0.0
  contact:
    name: TalkFlow Support
    email: support@talkflow.example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.talkflow.example.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        name:
          type: string
        avatar:
          type: string
          nullable: true
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  nullable: true
                phone:
                  type: string
                  nullable: true
                name:
                  type: string
                password:
                  type: string
                  format: password
              required: [name, password]
              anyOf:
                - required: [email]
                - required: [phone]
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  nullable: true
                phone:
                  type: string
                  nullable: true
                password:
                  type: string
                  format: password
              required: [password]
              anyOf:
                - required: [email]
                - required: [phone]
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Access token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
...
  /auth/verify:
    post:
      tags:
        - Auth
      summary: Verify email or phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  nullable: true
                phone:
                  type: string
                  nullable: true
                verificationCode:
                  type: string
              required: [verificationCode]
              anyOf:
                - required: [email]
                - required: [phone]
      responses:
        '200':
          description: Verification successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/send-verification:
    post:
      tags:
        - Auth
      summary: Send verification code via email or SMS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  nullable: true
                phone:
                  type: string
                  nullable: true
              anyOf:
                - required: [email]
                - required: [phone]
      responses:
        '200':
          description: Verification code sent successfully